Create a test for ed25519 KMS provider
