+++ src/kms/key-providers/index.ts
@@ -1,2 +1,3 @@
 export * from './bjj-provider';
+export * from './ed25519-provider';
 export * from './sec256k1-provider';
